+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/listener/RichPresenceCaretListener.kt
@@ -11,12 +11,13 @@ import kotlinx.coroutines.launch
 object RichPresenceCaretListener : CaretListener {
     override fun caretPositionChanged(event: CaretEvent) {
         val line = event.newPosition.line + 1
+        val column = event.newPosition.column + 1
         val discordService = DiscordService.getInstance()
         discordService.scope.launch(Dispatchers.Default) {
             discordService.modifyActivity { context ->
-                val project = context?.project?.get()
-                val editor = project?.let { FileEditorManager.getInstance(it) }?.selectedTextEditor
-                if (context?.file?.line != line && editor != null) {
+                val project = context?.project?.get() ?: return@modifyActivity context
+                if (context.file?.line != line || context.file.column != column) {
+                    val editor = FileEditorManager.getInstance(project).selectedTextEditor ?: return@modifyActivity context
                     ActivityContext.create(project, editor.virtualFile, editor)
                 } else {
                     context